refactor user
