+++ feature-libs/user/schematics/add-user/index.ts
@@ -1,20 +1,15 @@
 import {
   chain,
-  noop,
   Rule,
   SchematicContext,
   Tree,
 } from '@angular-devkit/schematics';
 import {
-  addLibraryFeature,
+  addFeatures,
   addPackageJsonDependenciesForLibrary,
-  CLI_USER_ACCOUNT_FEATURE,
-  CLI_USER_PROFILE_FEATURE,
+  analyzeCrossFeatureDependencies,
   LibraryOptions as SpartacusUserOptions,
   readPackageJson,
-  shouldAddFeature,
-  USER_ACCOUNT_SCHEMATICS_CONFIG,
-  USER_PROFILE_SCHEMATICS_CONFIG,
   validateSpartacusInstallation,
 } from '@spartacus/schematics';
 import { peerDependencies } from '../../package.json';
@@ -24,24 +19,11 @@ export function addUserFeatures(options: SpartacusUserOptions): Rule {
     const packageJson = readPackageJson(tree);
     validateSpartacusInstallation(packageJson);
 
+    const features = analyzeCrossFeatureDependencies(options.features ?? []);
+
     return chain([
+      addFeatures(options, features),
       addPackageJsonDependenciesForLibrary(peerDependencies, options),
-
-      shouldAddFeature(CLI_USER_ACCOUNT_FEATURE, options.features)
-        ? addAccountFeature(options)
-        : noop(),
-
-      shouldAddFeature(CLI_USER_PROFILE_FEATURE, options.features)
-        ? addProfileFeature(options)
-        : noop(),
     ]);
   };
-}
-
-function addAccountFeature(options: SpartacusUserOptions): Rule {
-  return addLibraryFeature(options, USER_ACCOUNT_SCHEMATICS_CONFIG);
-}
-
-function addProfileFeature(options: SpartacusUserOptions): Rule {
-  return addLibraryFeature(options, USER_PROFILE_SCHEMATICS_CONFIG);
 }
