+++ feature-libs/user/schematics/add-user/index_spec.ts
@@ -1,6 +1,5 @@
 /// <reference types="jest" />
 
-import { RunSchematicTaskOptions } from '@angular-devkit/schematics/tasks/run-schematic/options';
 import {
   SchematicTestRunner,
   UnitTestTree,
@@ -13,11 +12,9 @@ import { Schema as WorkspaceOptions } from '@schematics/angular/workspace/schema
 import {
   CLI_USER_ACCOUNT_FEATURE,
   CLI_USER_PROFILE_FEATURE,
-  LibraryOptions,
   LibraryOptions as SpartacusUserOptions,
   SpartacusOptions,
   SPARTACUS_SCHEMATICS,
-  SPARTACUS_USER,
 } from '@spartacus/schematics';
 import * as path from 'path';
 import { peerDependencies } from '../../package.json';
@@ -206,24 +203,9 @@ describe('Spartacus User schematics: ng-add', () => {
         });
       });
 
-      it('should run the proper installation tasks', async () => {
-        const tasks = schematicRunner.tasks
-          .filter((task) => task.name === 'run-schematic')
-          .map(
-            (task) => task.options as RunSchematicTaskOptions<LibraryOptions>
-          );
-        expect(tasks.length).toEqual(1);
-
-        const userTaskWithSubFeatures = tasks[0];
-        expect(userTaskWithSubFeatures).toBeTruthy();
-        expect(userTaskWithSubFeatures.name).toEqual('add-spartacus-library');
-        expect(userTaskWithSubFeatures.options).toHaveProperty(
-          'collection',
-          SPARTACUS_USER
-        );
-        expect(userTaskWithSubFeatures.options.options?.features).toEqual([
-          CLI_USER_ACCOUNT_FEATURE,
-        ]);
+      it('should have added User Account feature', async () => {
+        const featureModule = appTree.readContent(featureModulePath);
+        expect(featureModule).toMatchSnapshot();
       });
     });
