+++ projects/schematics/src/shared/utils/index.ts
@@ -1,9 +1,13 @@
 export * from './config-utils';
+export * from './dependency-utils';
+export * from './feature-utils';
 export * from './file-utils';
+export * from './graph-utils';
 export * from './lib-utils';
 export * from './module-file-utils';
 export * from './new-module-utils';
 export * from './package-utils';
 export * from './program';
 export * from './project-tsconfig-paths';
+export * from './schematics-config-utils';
 export * from './workspace-utils';
